# module/OAuth/App/service-rest/src/api-spec/authorization-request/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "OAuth 2.0 Provider API"
  },
  "servers": [
    {
      "url": "/oauth/api"
    }
  ],
  "tags": [
    {
      "name": "authorization-request: entity",
      "description": "Entity operations (CRUD)"
    },
    {
      "name": "authorization-request: search",
      "description": "AuthorizationRequest search operations"
    }
  ],
  "paths": {
    "/authorization-request/{authRequestId}": {
      "get": {
        "tags": ["authorization-request: entity"],
        "summary": "Get record by ID",
        "description": "",
        "operationId": "getRecordById-authorization-request",
        "parameters": [
          {
            "name": "authRequestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizationRequestDto"
                }
              }
            }
          }
        }
      }
    },
    "/authorization-request/search/{searchId}/resultset-size": {
      "get": {
        "tags": ["authorization-request: search"],
        "summary": "Get size of a resultset for a search request",
        "description": "",
        "operationId": "getSearchResultsetSize",
        "parameters": [
          {
            "name": "searchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Cache-Control",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/authorization-request/search/{searchId}/resultset/paged-by-{pageSize}/{page}": {
      "get": {
        "tags": ["authorization-request: search"],
        "summary": "Get paged resultset for a search request",
        "description": "",
        "operationId": "getResultsetPaged",
        "parameters": [
          {
            "name": "searchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "pattern": "\\d+"
            }
          },
          {
            "name": "page",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Cache-Control",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthorizationRequestDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/authorization-request/{authRequestId}/block": {
      "post": {
        "tags": ["authorization-request: entity"],
        "summary": "Block authorization request",
        "description": "",
        "operationId": "blockAuthorizationRequest",
        "parameters": [
          {
            "name": "authRequestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
    "/authorization-request/search": {
      "post": {
        "tags": ["authorization-request: search"],
        "summary": "Post search request",
        "description": "",
        "operationId": "postSearch",
        "parameters": [
          {
            "name": "Extended-Response",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Cache-Control",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json;charset=utf-8": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequestDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/authorization-request/search/{searchId}/resultset": {
      "get": {
        "tags": ["authorization-request: search"],
        "summary": "Get resultset for a search request",
        "description": "",
        "operationId": "getResultset",
        "parameters": [
          {
            "name": "searchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Cache-Control",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthorizationRequestDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/authorization-request/search/{searchId}": {
      "get": {
        "tags": ["authorization-request: search"],
        "summary": "Get search request by ID",
        "description": "",
        "operationId": "getSearchRequest",
        "parameters": [
          {
            "name": "searchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/SearchRequestDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ColumnSortConfigurationDto": {
        "type": "object",
        "properties": {
          "columnName": {
            "type": "string"
          },
          "sortOrder": {
            "type": "string"
          }
        }
      },
      "OptionDtoInteger": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "integer"
          }
        }
      },
      "OptionDtoString": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "AuthorizationRequestDto": {
        "type": "object",
        "properties": {
          "authRequestId": {
            "type": "integer"
          },
          "authorizationCode": {
            "type": "string"
          },
          "dateIns": {
            "type": "string",
            "format": "date-time"
          },
          "operator": {
            "$ref": "#/components/schemas/OptionDtoInteger"
          },
          "operatorLogin": {
            "type": "string"
          },
          "tokenId": {
            "type": "string"
          },
          "tokenDateIns": {
            "type": "string",
            "format": "date-time"
          },
          "redirectUri": {
            "type": "string"
          },
          "client": {
            "$ref": "#/components/schemas/OptionDtoString"
          },
          "isBlocked": {
            "type": "boolean"
          }
        }
      },
      "AuthorizationRequestSearchDto": {
        "type": "object",
        "properties": {
          "authRequestId": {
            "type": "integer"
          },
          "operatorId": {
            "type": "integer"
          },
          "clientId": {
            "type": "string"
          },
          "maxRowCount": {
            "type": "integer"
          }
        }
      },
      "SearchRequestDto": {
        "type": "object",
        "properties": {
          "template": {
            "$ref": "#/components/schemas/AuthorizationRequestSearchDto"
          },
          "listSortConfiguration": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColumnSortConfigurationDto"
            }
          }
        }
      }
    }
  }
}